+++ crates/cheatnet/src/cheatcodes/roll.rs
@@ -1,13 +1,43 @@
+use crate::state::{CheatStatus, CheatTarget};
 use crate::CheatnetState;
 use cairo_felt::Felt252;
-use starknet_api::core::ContractAddress;
 
 impl CheatnetState {
-    pub fn start_roll(&mut self, contract_address: ContractAddress, block_number: Felt252) {
-        self.rolled_contracts.insert(contract_address, block_number);
+    pub fn start_roll(&mut self, target: CheatTarget, block_number: Felt252) {
+        match target {
+            CheatTarget::All => {
+                self.global_roll = Some(block_number);
+                self.rolled_contracts.clear();
+            }
+            CheatTarget::One(contract_address) => {
+                self.rolled_contracts
+                    .insert(contract_address, CheatStatus::Cheated(block_number));
+            }
+            CheatTarget::Multiple(contracts) => {
+                for contract_address in contracts {
+                    self.rolled_contracts
+                        .insert(contract_address, CheatStatus::Cheated(block_number.clone()));
+                }
+            }
+        }
     }
 
-    pub fn stop_roll(&mut self, contract_address: ContractAddress) {
-        self.rolled_contracts.remove(&contract_address);
+    pub fn stop_roll(&mut self, target: CheatTarget) {
+        match target {
+            CheatTarget::All => {
+                self.global_roll = None;
+                self.rolled_contracts.clear();
+            }
+            CheatTarget::One(contract_address) => {
+                self.rolled_contracts
+                    .insert(contract_address, CheatStatus::Uncheated);
+            }
+            CheatTarget::Multiple(contracts) => {
+                for contract_address in contracts {
+                    self.rolled_contracts
+                        .insert(contract_address, CheatStatus::Uncheated);
+                }
+            }
+        }
     }
 }