+++ crates/cheatnet/src/execution/execution_info.rs
@@ -18,8 +18,8 @@ fn get_cheated_block_info_ptr(
 
     let mut new_block_info = original_block_info.to_owned();
 
-    if let Some(rolled_number) = cheatnet_state.rolled_contracts.get(contract_address) {
-        new_block_info[0] = MaybeRelocatable::Int(rolled_number.clone());
+    if let Some(rolled_number) = cheatnet_state.get_cheated_block_number(contract_address) {
+        new_block_info[0] = MaybeRelocatable::Int(rolled_number);
     };
 
     if let Some(warped_timestamp) = cheatnet_state.warped_contracts.get(contract_address) {