+++ crates/cheatnet/tests/starknet/cheat_fork.rs
@@ -2,6 +2,7 @@ use crate::common::felt_selector_from_name;
 use crate::common::state::{create_cheatnet_state, create_fork_cached_state};
 use cairo_felt::Felt252;
 use cheatnet::rpc::{call_contract, CallContractResult};
+use cheatnet::state::CheatTarget;
 use conversions::StarknetConversions;
 use num_bigint::BigUint;
 use starknet_api::core::ContractAddress;
@@ -91,7 +92,7 @@ fn roll_cairo0_contract(selector: &str) {
     };
     let block_number = &ret_data[0];
 
-    cheatnet_state.start_roll(contract_address, Felt252::from(123));
+    cheatnet_state.start_roll(CheatTarget::One(contract_address), Felt252::from(123));
 
     let output = call_contract(
         &mut blockifier_state,
@@ -106,7 +107,7 @@ fn roll_cairo0_contract(selector: &str) {
     };
     let rolled_block_number = &ret_data[0];
 
-    cheatnet_state.stop_roll(contract_address);
+    cheatnet_state.stop_roll(CheatTarget::One(contract_address));
 
     let output = call_contract(
         &mut blockifier_state,