+++ crates/forge/src/test_execution_syscall_handler.rs
@@ -25,7 +25,7 @@ use cheatnet::cheatcodes::deploy::{deploy, deploy_at, DeployCallPayload};
 use cheatnet::cheatcodes::{CheatcodeError, EnhancedHintError};
 use cheatnet::execution::cheatable_syscall_handler::CheatableSyscallHandler;
 use cheatnet::rpc::{call_contract, CallContractFailure, CallContractOutput, CallContractResult};
-use cheatnet::state::{BlockifierState, CheatnetState};
+use cheatnet::state::{BlockifierState, CheatTarget, CheatnetState};
 use conversions::StarknetConversions;
 use num_traits::{One, ToPrimitive};
 use scarb_artifacts::StarknetContractArtifacts;
@@ -241,20 +241,20 @@ impl TestExecutionSyscallHandler<'_, '_> {
 
         match selector {
             "start_roll" => {
-                let contract_address = inputs[0].to_contract_address();
-                let value = inputs[1].clone();
+                let target = deserialize_cheat_target(&inputs[..inputs.len() - 1]);
+                let block_number = inputs.last().unwrap().clone();
                 self.contract_execution_syscall_handler
                     .cheatable_syscall_handler
                     .cheatnet_state
-                    .start_roll(contract_address, value);
+                    .start_roll(target, block_number);
                 Ok(())
             }
             "stop_roll" => {
-                let contract_address = inputs[0].to_contract_address();
+                let target = deserialize_cheat_target(&inputs);
                 self.contract_execution_syscall_handler
                     .cheatable_syscall_handler
                     .cheatnet_state
-                    .stop_roll(contract_address);
+                    .stop_roll(target);
                 Ok(())
             }
             "start_warp" => {
@@ -692,6 +692,22 @@ fn handle_deploy_result(
             Ok(())
         }
         Err(CheatcodeError::Unrecoverable(err)) => Err(err),
+    }
+}
+
+fn deserialize_cheat_target(inputs: &[Felt252]) -> CheatTarget {
+    // First element encodes the variant of CheatTarget
+    match inputs[0].to_u8() {
+        Some(0) => CheatTarget::All,
+        Some(1) => CheatTarget::One(inputs[1].to_contract_address()),
+        Some(2) => {
+            let contract_addresses: Vec<_> = inputs[2..]
+                .iter()
+                .map(Felt252::to_contract_address)
+                .collect();
+            CheatTarget::Multiple(contract_addresses)
+        }
+        _ => panic!("Invalid CheatTarget variant"),
     }
 }